requirements.txt
flask
flask_sqlalchemy
mediapipe
opencv-python-headless
numpy
pillow
pytz
gunicorn
